--- src/main/kotlin/it/pagopa/ecommerce/transactions/scheduler/deadletter/responseBuildersUtil.kt
@@ -1,52 +1,0 @@
-package it.pagopa.ecommerce.transactions.scheduler.deadletter
-
-import it.pagopa.ecommerce.commons.documents.v2.TransactionActivatedData
-import it.pagopa.ecommerce.commons.documents.v2.TransactionAuthorizationCompletedData
-import it.pagopa.ecommerce.commons.documents.v2.TransactionAuthorizationRequestData
-import it.pagopa.ecommerce.commons.documents.v2.authorization.*
-import it.pagopa.ecommerce.commons.domain.v2.pojos.*
-import it.pagopa.ecommerce.commons.utils.v2.TransactionUtils.getTransactionFee
-import java.util.*
-
-fun getTransactionActivatedData(baseTransaction: BaseTransaction): TransactionActivatedData? =
-    if (baseTransaction is BaseTransactionWithPaymentToken) {
-        baseTransaction.transactionActivatedData
-    } else {
-        null
-    }
-
-fun getTransactionFees(baseTransaction: BaseTransaction): Optional<Int> =
-    when (baseTransaction) {
-        is BaseTransactionExpired -> getTransactionFee(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithClosureError ->
-            getTransactionFee(baseTransaction.transactionAtPreviousState)
-        else -> getTransactionFee(baseTransaction)
-    }
-
-fun getTransactionAuthRequestedData(
-    baseTransaction: BaseTransaction
-): TransactionAuthorizationRequestData? =
-    when (baseTransaction) {
-        is BaseTransactionExpired ->
-            getTransactionAuthRequestedData(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithClosureError ->
-            getTransactionAuthRequestedData(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithRequestedAuthorization ->
-            baseTransaction.transactionAuthorizationRequestData
-        else -> null
-    }
-
-fun getTransactionAuthCompletedData(
-    baseTransaction: BaseTransaction
-): TransactionAuthorizationCompletedData? =
-    when (baseTransaction) {
-        is BaseTransactionExpired ->
-            getTransactionAuthCompletedData(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithClosureError ->
-            getTransactionAuthCompletedData(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithRefundRequested ->
-            getTransactionAuthCompletedData(baseTransaction.transactionAtPreviousState)
-        is BaseTransactionWithCompletedAuthorization ->
-            baseTransaction.transactionAuthorizationCompletedData
-        else -> null
-    }
